requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
pydantic>=2.6
pyyaml>=6.0
tenacity>=8.2
pytest>=8.0
pytest-asyncio>=0.23
